refactor(derivation): split new logic into verify.go and reorg.go
Extract newly added functions into dedicated files for clarity:
- verify.go: rollbackLocalChain, verifyBatchRoots, verifyBlockContext
- reorg.go: detectReorg, handleL1Reorg, recordL1Blocks

Existing batch parsing code stays in derivation.go to keep the diff
scoped to this PR's changes only. No logic changes — pure file split.

Made-with: Cursor
