+++ node/derivation/derivation.go
@@ -25,7 +25,6 @@ import (
 	"morph-l2/bindings/bindings"
 	"morph-l2/bindings/predeploys"
 	nodecommon "morph-l2/node/common"
-	"morph-l2/node/db"
 	"morph-l2/node/sync"
 	"morph-l2/node/types"
 )
@@ -364,196 +363,6 @@ func (d *Derivation) derivationBlock(ctx context.Context) {
 	d.logger.Info("write latest derivation l1 height success", "l1BlockNumber", end)
 }
 
-// detectReorg checks recent L1 blocks for hash mismatches indicating a reorg.
-// Returns the L1 height where reorg was first detected, or nil if no reorg.
-//
-// Optimization: checks the newest saved block first. If it matches, there is
-// no reorg (1 RPC call in the common case). Only when the newest block
-// mismatches does it do a full oldest-to-newest scan to find the earliest
-// divergence point.
-func (d *Derivation) detectReorg(ctx context.Context) (*uint64, error) {
-	latestDerivation := d.db.ReadLatestDerivationL1Height()
-	if latestDerivation == nil {
-		return nil, nil
-	}
-
-	checkFrom := d.startHeight
-	if *latestDerivation > d.reorgCheckDepth && (*latestDerivation-d.reorgCheckDepth) > checkFrom {
-		checkFrom = *latestDerivation - d.reorgCheckDepth
-	}
-
-	savedBlocks := d.db.ReadDerivationL1BlockRange(checkFrom, *latestDerivation)
-	if len(savedBlocks) == 0 {
-		return nil, nil
-	}
-
-	// Fast path: check the newest block first. If it matches, no reorg occurred.
-	newest := savedBlocks[len(savedBlocks)-1]
-	newestHeader, err := d.l1Client.HeaderByNumber(ctx, big.NewInt(int64(newest.Number)))
-	if err != nil {
-		return nil, fmt.Errorf("failed to get L1 header at %d: %w", newest.Number, err)
-	}
-	if newestHeader.Hash() == common.BytesToHash(newest.Hash[:]) {
-		return nil, nil
-	}
-
-	// Slow path: reorg detected. Scan oldest-to-newest to find the earliest divergence.
-	for i := 0; i < len(savedBlocks); i++ {
-		block := savedBlocks[i]
-		header, err := d.l1Client.HeaderByNumber(ctx, big.NewInt(int64(block.Number)))
-		if err != nil {
-			return nil, fmt.Errorf("failed to get L1 header at %d: %w", block.Number, err)
-		}
-		savedHash := common.BytesToHash(block.Hash[:])
-		if header.Hash() != savedHash {
-			d.logger.Info("L1 block hash mismatch detected",
-				"height", block.Number,
-				"savedHash", savedHash.Hex(),
-				"currentHash", header.Hash().Hex(),
-			)
-			return &block.Number, nil
-		}
-	}
-	return nil, nil
-}
-
-// handleL1Reorg handles an L1 reorg detected at the given L1 height.
-// It only cleans up derivation DB state and resets the derivation L1 height
-// so the next derivation loop re-processes from the reorg point.
-//
-// L1 reorg does NOT directly trigger an L2 rollback — in most cases the same
-// batch tx gets re-included in a new L1 block with identical content, so L2
-// blocks remain valid. The normal derivation loop will re-fetch batches and
-// run verifyBlockContext / verifyBatchRoots; only if those comparisons fail
-// will an L2 rollback be triggered through rollbackLocalChain.
-func (d *Derivation) handleL1Reorg(reorgAtL1Height uint64) error {
-	d.logger.Info("L1 reorg detected, cleaning DB records and restarting derivation from reorg point",
-		"reorgAtL1Height", reorgAtL1Height)
-
-	d.db.DeleteDerivationL1BlocksFrom(reorgAtL1Height)
-
-	if reorgAtL1Height > d.startHeight {
-		d.db.WriteLatestDerivationL1Height(reorgAtL1Height - 1)
-	} else {
-		// Reorg at or before startHeight — reset so next loop starts from startHeight.
-		if d.startHeight > 0 {
-			d.db.WriteLatestDerivationL1Height(d.startHeight - 1)
-		} else {
-			d.db.WriteLatestDerivationL1Height(0)
-		}
-	}
-
-	return nil
-}
-
-// rollbackLocalChain rolls back the local L2 chain to the specified block number.
-// This is only triggered when batch data comparison fails — i.e. the local L2 block
-// does not match the L1 batch data (block context mismatch or state root mismatch).
-// After rollback, the caller re-derives blocks using L1 batch data as source of truth.
-func (d *Derivation) rollbackLocalChain(targetBlockNumber uint64) error {
-	d.logger.Error("L2 chain rollback not yet implemented",
-		"targetBlockNumber", targetBlockNumber)
-
-	// TODO: Implement actual rollback via geth SetHead engine API:
-	//  1. Expose SetL2Head(number uint64) in go-ethereum/eth/catalyst/l2_api.go
-	//  2. Add SetHead method to go-ethereum/ethclient/authclient
-	//  3. Add SetHead method to node/types/retryable_client.go
-	//  4. Call d.l2Client.SetHead(d.ctx, targetBlockNumber)
-	return fmt.Errorf("rollback not implemented yet, target=%d", targetBlockNumber)
-}
-
-// verifyBatchRoots verifies that the local state root and withdrawal root match the L1 batch data.
-func (d *Derivation) verifyBatchRoots(batchInfo *BatchInfo, lastHeader *eth.Header) error {
-	withdrawalRoot, err := d.L2ToL1MessagePasser.MessageRoot(&bind.CallOpts{
-		BlockNumber: lastHeader.Number,
-	})
-	if err != nil {
-		return fmt.Errorf("get withdrawal root failed: %w", err)
-	}
-
-	rootMismatch := !bytes.Equal(lastHeader.Root.Bytes(), batchInfo.root.Bytes())
-	withdrawalMismatch := !bytes.Equal(withdrawalRoot[:], batchInfo.withdrawalRoot.Bytes())
-
-	if rootMismatch || withdrawalMismatch {
-		// Check if should skip validation during upgrade transition
-		if d.switchTime > 0 {
-			beforeSwitch := lastHeader.Time < d.switchTime
-			if (beforeSwitch && !d.useZktrie) || (!beforeSwitch && d.useZktrie) {
-				d.logger.Info("Root validation skipped during upgrade transition",
-					"originStateRootHash", batchInfo.root,
-					"deriveStateRootHash", lastHeader.Root.Hex(),
-					"blockTimestamp", lastHeader.Time,
-					"switchTime", d.switchTime,
-					"useZktrie", d.useZktrie,
-				)
-				return nil
-			}
-		}
-		return fmt.Errorf("root mismatch: stateRoot(l1=%s, local=%s) withdrawalRoot(l1=%s, local=%s)",
-			batchInfo.root.Hex(), lastHeader.Root.Hex(),
-			batchInfo.withdrawalRoot.Hex(), common.BytesToHash(withdrawalRoot[:]).Hex())
-	}
-	return nil
-}
-
-// verifyBlockContext compares a local L2 block header against the batch block context from L1.
-func (d *Derivation) verifyBlockContext(localHeader *eth.Header, blockData *BlockContext) error {
-	if localHeader.Time != blockData.Timestamp {
-		return fmt.Errorf("timestamp mismatch at block %d: local=%d, batch=%d",
-			blockData.Number, localHeader.Time, blockData.Timestamp)
-	}
-	if localHeader.GasLimit != blockData.GasLimit {
-		return fmt.Errorf("gasLimit mismatch at block %d: local=%d, batch=%d",
-			blockData.Number, localHeader.GasLimit, blockData.GasLimit)
-	}
-	switch {
-	case blockData.BaseFee != nil && localHeader.BaseFee != nil:
-		if localHeader.BaseFee.Cmp(blockData.BaseFee) != 0 {
-			return fmt.Errorf("baseFee mismatch at block %d: local=%s, batch=%s",
-				blockData.Number, localHeader.BaseFee.String(), blockData.BaseFee.String())
-		}
-	case blockData.BaseFee == nil && localHeader.BaseFee == nil:
-		// Both nil — pre-EIP-1559 or legacy batch format, OK.
-	default:
-		// One side has BaseFee, the other doesn't — structural inconsistency.
-		return fmt.Errorf("baseFee nil mismatch at block %d: local=%v, batch=%v",
-			blockData.Number, localHeader.BaseFee, blockData.BaseFee)
-	}
-	// Batch internal consistency check: txsNum in the block context should match the
-	// actual number of transactions assembled in SafeL2Data (L1 messages + L2 txs).
-	// This catches batch parsing/corruption issues, not local-vs-L1 divergence.
-	// Local-vs-L1 transaction divergence is covered by state root verification
-	// in verifyBatchRoots (different txs → different state root).
-	if blockData.SafeL2Data != nil {
-		batchTxCount := len(blockData.SafeL2Data.Transactions)
-		if batchTxCount != int(blockData.txsNum) {
-			return fmt.Errorf("batch internal tx count inconsistency at block %d: blockContext.txsNum=%d, safeL2Data.Transactions=%d",
-				blockData.Number, blockData.txsNum, batchTxCount)
-		}
-	}
-	return nil
-}
-
-// recordL1Blocks saves L1 block hashes for reorg detection.
-// Returns an error if any header fetch fails — the caller must not advance
-// derivation height to avoid permanent gaps in L1 block hash tracking.
-func (d *Derivation) recordL1Blocks(ctx context.Context, from, to uint64) error {
-	for h := from; h <= to; h++ {
-		header, err := d.l1Client.HeaderByNumber(ctx, big.NewInt(int64(h)))
-		if err != nil {
-			return fmt.Errorf("failed to get L1 header at %d: %w", h, err)
-		}
-
-		var hashBytes [32]byte
-		copy(hashBytes[:], header.Hash().Bytes())
-
-		d.db.WriteDerivationL1Block(&db.DerivationL1Block{
-			Number: h,
-			Hash:   hashBytes,
-		})
-	}
-	return nil
-}
 
 func (d *Derivation) fetchRollupLog(ctx context.Context, from, to uint64) ([]eth.Log, error) {
 	query := ethereum.FilterQuery{
